+++ rtic-macros/src/codegen/bindings/riscv_slic.rs
@@ -186,7 +186,7 @@ pub fn check_stack_overflow_before_init(
         if stack_start > ebss {
             // No flip-link usage, check the SP for overflow.
             if rtic::export::read_sp() <= ebss {
-                panic!("Stack overflow after allocating executors");
+                ::core::panic!("Stack overflow after allocating executors");
             }
         }
     )]